refactor(parser): split grammar/mod.rs into focused submodules
Pure code-motion split of the oversized src/parser/grammar/mod.rs
(1506 lines) per issue #391 Unit U4. mod.rs retained only top-level
orchestration (parse_magic_rule, strip_optional_x_operator) plus mod
declarations and re-exports; extracted five new sibling modules:

- offset.rs: pointer_specifier_to_type, parse_indirect_offset,
  parse_offset, parse_rule_offset
- operator.rs: parse_operator
- type_and_operator.rs: parse_name_or_use_meta,
  parse_type_and_operator, parse_type
- rule_value.rs: parse_string_family_value, parse_bare_string_value
- lines.rs: parse_message, parse_strength_directive,
  is_strength_directive, parse_comment, is_empty_line,
  is_comment_line, has_continuation

No logic or behavior changes -- all existing call sites and grammar
unit tests keep working unchanged via re-exports at the grammar
module level. mod.rs is now 284 lines; every new file is under 400
lines, well within the 500-600 line ceiling.
