+++ src/parser/grammar/mod.rs
@@ -5,1032 +5,57 @@
 //!
 //! This module implements the parsing logic for magic file syntax, converting
 //! text-based magic rules into the AST representation defined in ast.rs.
+//!
+//! Submodules split out focused parsing concerns (issue #391 Unit U4):
+//! offset specifications (`offset`), comparison operators (`operator`),
+//! type+operator+transform parsing (`type_and_operator`), string-family
+//! bareword value parsing (`rule_value`), and line-level helpers
+//! (`lines`). This top-level module retains the main dispatch entry point,
+//! [`parse_magic_rule`].
 
 use nom::{
-    IResult, Parser,
-    branch::alt,
-    bytes::complete::{tag, take_while},
-    character::complete::{char, multispace0, one_of},
-    combinator::opt,
-    error::Error as NomError,
-    multi::many0,
-    sequence::preceded,
+    IResult, Parser, character::complete::multispace0, combinator::opt, error::Error as NomError,
 };
 
-use log::warn;
-
-use crate::parser::ast::{
-    Endianness, IndirectAdjustmentOp, MagicRule, MetaType, OffsetSpec, Operator, StrengthModifier,
-    TypeKind, Value,
-};
+use crate::parser::ast::{MagicRule, MetaType, Operator, TypeKind, Value};
+#[cfg(test)]
+use crate::parser::ast::{OffsetSpec, StrengthModifier};
 
 mod getstr;
+mod lines;
 mod numbers;
+mod offset;
+mod operator;
+mod rule_value;
+mod type_and_operator;
 mod type_suffix;
 mod value;
 
+pub use lines::{
+    has_continuation, is_comment_line, is_empty_line, is_strength_directive, parse_comment,
+    parse_message, parse_strength_directive,
+};
 pub use numbers::parse_number;
+// `parse_offset` has no non-test caller outside `offset.rs` itself (it is
+// invoked internally by `parse_rule_offset`); the re-export exists so
+// `crate::parser::grammar::parse_offset` keeps resolving for grammar unit
+// tests (`use super::*` in `grammar/tests/`), matching its original
+// directly-defined-in-mod.rs visibility.
+#[allow(unused_imports)]
+pub use offset::{parse_offset, parse_rule_offset};
+pub use operator::parse_operator;
+// `parse_type` is a standalone helper exercised by grammar unit tests only
+// (see its `#[allow(dead_code)]` at the definition site).
+#[allow(unused_imports)]
+pub use type_and_operator::{parse_type, parse_type_and_operator};
 pub use value::parse_value;
 
-use numbers::parse_decimal_number;
+use rule_value::parse_string_family_value;
+
 #[cfg(test)]
-use numbers::parse_hex_number;
-use type_suffix::{
-    parse_attached_operator, parse_pstring_suffix, parse_regex_suffix, parse_search_suffix,
-    parse_value_transform,
-};
+use numbers::{parse_decimal_number, parse_hex_number};
 #[cfg(test)]
 use value::{parse_escape_sequence, parse_hex_bytes, parse_numeric_value, parse_quoted_string};
-
-/// Map a single-character pointer specifier to its `TypeKind` and `Endianness`.
-///
-/// GNU `file` semantics: lowercase = little-endian, uppercase = big-endian.
-/// Numeric pointer types are signed by default per GOTCHAS S6.3.
-///
-/// | Specifier | Width  | Endianness    |
-/// |-----------|--------|---------------|
-/// | `b`       | 1 byte | Little-endian |
-/// | `B`       | 1 byte | Big-endian    |
-/// | `s`       | 2 byte | Little-endian |
-/// | `S`       | 2 byte | Big-endian    |
-/// | `l`       | 4 byte | Little-endian |
-/// | `L`       | 4 byte | Big-endian    |
-/// | `q`       | 8 byte | Little-endian |
-/// | `Q`       | 8 byte | Big-endian    |
-fn pointer_specifier_to_type(spec: char) -> Option<(TypeKind, Endianness)> {
-    match spec {
-        'b' => Some((TypeKind::Byte { signed: true }, Endianness::Little)),
-        'B' => Some((TypeKind::Byte { signed: true }, Endianness::Big)),
-        's' => Some((
-            TypeKind::Short {
-                endian: Endianness::Little,
-                signed: true,
-            },
-            Endianness::Little,
-        )),
-        'S' => Some((
-            TypeKind::Short {
-                endian: Endianness::Big,
-                signed: true,
-            },
-            Endianness::Big,
-        )),
-        'l' => Some((
-            TypeKind::Long {
-                endian: Endianness::Little,
-                signed: true,
-            },
-            Endianness::Little,
-        )),
-        'L' => Some((
-            TypeKind::Long {
-                endian: Endianness::Big,
-                signed: true,
-            },
-            Endianness::Big,
-        )),
-        'q' => Some((
-            TypeKind::Quad {
-                endian: Endianness::Little,
-                signed: true,
-            },
-            Endianness::Little,
-        )),
-        'Q' => Some((
-            TypeKind::Quad {
-                endian: Endianness::Big,
-                signed: true,
-            },
-            Endianness::Big,
-        )),
-        // `i` and `I` are magic(5) "ID3 variable-byte int" pointer
-        // specifiers used in audio:308 for ID3 frame size decoding.
-        // We parse them so the magic file loads, but for now treat
-        // them as plain 32-bit longs with the corresponding endianness
-        // -- real ID3 7-bit-per-byte decoding is a follow-up. Tracked
-        // separately as a parsing-vs-semantics gap. The bodies match
-        // `l`/`L` exactly today; clippy::match_same_arms is allowed
-        // because the arms are intentionally distinct entry points
-        // that future ID3-decoding work will diverge.
-        #[allow(clippy::match_same_arms)]
-        'i' => Some((
-            TypeKind::Long {
-                endian: Endianness::Little,
-                signed: true,
-            },
-            Endianness::Little,
-        )),
-        #[allow(clippy::match_same_arms)]
-        'I' => Some((
-            TypeKind::Long {
-                endian: Endianness::Big,
-                signed: true,
-            },
-            Endianness::Big,
-        )),
-        _ => None,
-    }
-}
-
-/// Parse an indirect offset specification with optional arithmetic.
-///
-/// Accepts these forms:
-///
-/// - `(base.type)` — no adjustment
-/// - `(base.type+N)` / `(base.type-N)` — additive (canonical magic(5))
-/// - `(base.type*N)` / `(base.type/N)` / `(base.type%N)` — multiplicative
-/// - `(base.type&N)` / `(base.type|N)` / `(base.type^N)` — bitwise
-/// - `(base.type)+N` / `(base.type)-N` — additive outside the parens
-///   (backwards-compatible alternate form; only `+`/`-` are accepted here)
-///
-/// Only one adjustment form may be used per rule; combinations like
-/// `(19.b-1)+2` or `(0x200.s*2)+4` are not permitted. Subtraction is
-/// represented as [`IndirectAdjustmentOp::Add`] with a negative
-/// `adjustment`.
-fn parse_indirect_offset(input: &str) -> IResult<&str, OffsetSpec> {
-    // Inside-paren adjustment supports the full magic(5) operator set.
-    // Returns `Some((op, value))` when an operator+operand was consumed.
-    //
-    // Operands may optionally be wrapped in their own parentheses, e.g.
-    // `(0x10.l+(-4))` is equivalent to `(0x10.l-4)`. GNU `file` magic
-    // files use this form when a sign character would otherwise be
-    // ambiguous with the operator (e.g., `+-4`); the parens make the
-    // grouping explicit.
-    fn parse_operand(input: &str) -> IResult<&str, i64> {
-        if let Some(rest) = input.strip_prefix('(') {
-            let (rest, n) = parse_number(rest)?;
-            let (rest, _) = char(')')(rest)?;
-            Ok((rest, n))
-        } else {
-            parse_number(input)
-        }
-    }
-    fn parse_inside_adjustment(input: &str) -> IResult<&str, Option<(IndirectAdjustmentOp, i64)>> {
-        // Subtraction is folded into Add with a negated operand so the
-        // evaluator does not need a dedicated Sub variant.
-        if let Some(rest) = input.strip_prefix('+') {
-            let (rest, n) = parse_operand(rest)?;
-            Ok((rest, Some((IndirectAdjustmentOp::Add, n))))
-        } else if input.starts_with('-') {
-            let (rest, n) = parse_number(input)?;
-            Ok((rest, Some((IndirectAdjustmentOp::Add, n))))
-        } else if let Some(rest) = input.strip_prefix('*') {
-            let (rest, n) = parse_operand(rest)?;
-            Ok((rest, Some((IndirectAdjustmentOp::Mul, n))))
-        } else if let Some(rest) = input.strip_prefix('/') {
-            let (rest, n) = parse_operand(rest)?;
-            Ok((rest, Some((IndirectAdjustmentOp::Div, n))))
-        } else if let Some(rest) = input.strip_prefix('%') {
-            let (rest, n) = parse_operand(rest)?;
-            Ok((rest, Some((IndirectAdjustmentOp::Mod, n))))
-        } else if let Some(rest) = input.strip_prefix('&') {
-            let (rest, n) = parse_operand(rest)?;
-            Ok((rest, Some((IndirectAdjustmentOp::And, n))))
-        } else if let Some(rest) = input.strip_prefix('|') {
-            let (rest, n) = parse_operand(rest)?;
-            Ok((rest, Some((IndirectAdjustmentOp::Or, n))))
-        } else if let Some(rest) = input.strip_prefix('^') {
-            let (rest, n) = parse_operand(rest)?;
-            Ok((rest, Some((IndirectAdjustmentOp::Xor, n))))
-        } else {
-            Ok((input, None))
-        }
-    }
-
-    // Outside-paren adjustment: only `+`/`-` are accepted (legacy form).
-    fn parse_outside_adjustment(input: &str) -> IResult<&str, Option<i64>> {
-        if let Some(rest) = input.strip_prefix('+') {
-            let (rest, n) = parse_number(rest)?;
-            Ok((rest, Some(n)))
-        } else if input.starts_with('-') {
-            let (rest, n) = parse_number(input)?;
-            Ok((rest, Some(n)))
-        } else {
-            Ok((input, None))
-        }
-    }
-
-    let (input, _) = char('(')(input)?;
-
-    // magic(5) lets the indirect base itself be relative to the current
-    // anchor: `(&N.X)` means "read pointer at anchor + N". Detect the
-    // optional leading `&` and record the flag; the rest of the parser
-    // handles the numeric base offset uniformly.
-    let (input, base_relative) = if let Some(rest) = input.strip_prefix('&') {
-        (rest, true)
-    } else {
-        (input, false)
-    };
-
-    let (input, base_offset) = parse_number(input)?;
-    // magic(5) canonical separator is `.`. `/usr/share/file/magic/msdos`
-    // line 638 uses `,` -- a known typo that GNU `file` warns about
-    // but tolerates ("No current entry for continuation"). Accept
-    // either character so the magic file loads, but emit a warn! when
-    // the comma path is taken so users see the typo at default log
-    // levels (matching GNU `file`'s diagnostic posture).
-    let (input, sep) = one_of(".,").parse(input)?;
-    if sep == ',' {
-        warn!(
-            "Indirect offset uses ',' as separator (magic(5) requires '.'); \
-             accepting for GNU `file` typo-tolerance compatibility"
-        );
-    }
-    let (input, spec_char) = one_of("bBsSlLqQiI")(input)?;
-
-    let (pointer_type, endian) = pointer_specifier_to_type(spec_char)
-        .ok_or_else(|| nom::Err::Error(NomError::new(input, nom::error::ErrorKind::OneOf)))?;
-
-    let (input, inside) = parse_inside_adjustment(input)?;
-    let (input, _) = char(')')(input)?;
-
-    // Fall back to outside-paren adjustment if no inside form was present.
-    let (input, adjustment_op, adjustment) = if let Some((op, n)) = inside {
-        (input, op, n)
-    } else {
-        let (input, outside) = parse_outside_adjustment(input)?;
-        (input, IndirectAdjustmentOp::Add, outside.unwrap_or(0))
-    };
-
-    Ok((
-        input,
-        OffsetSpec::Indirect {
-            base_offset,
-            base_relative,
-            pointer_type,
-            adjustment,
-            adjustment_op,
-            result_relative: false,
-            endian,
-        },
-    ))
-}
-
-/// Parse an offset specification (absolute or indirect)
-///
-/// Supports:
-/// - Absolute offsets: decimal and hexadecimal, positive and negative
-/// - Indirect offsets: `(base.type)` or `(base.type)+adj` syntax
-///
-/// # Examples
-///
-/// ```ignore
-/// use libmagic_rs::parser::grammar::parse_offset;
-/// use libmagic_rs::parser::ast::{Endianness, OffsetSpec, TypeKind};
-///
-/// // Absolute offsets
-/// assert_eq!(parse_offset("0"), Ok(("", OffsetSpec::Absolute(0))));
-/// assert_eq!(parse_offset("123"), Ok(("", OffsetSpec::Absolute(123))));
-/// assert_eq!(parse_offset("0x10"), Ok(("", OffsetSpec::Absolute(16))));
-/// assert_eq!(parse_offset("-4"), Ok(("", OffsetSpec::Absolute(-4))));
-/// assert_eq!(parse_offset("-0xFF"), Ok(("", OffsetSpec::Absolute(-255))));
-///
-/// // Indirect offset (lowercase = little-endian, signed by default)
-/// assert_eq!(
-///     parse_offset("(0x3c.l)"),
-///     Ok(("", OffsetSpec::Indirect {
-///         base_offset: 0x3c,
-///         pointer_type: TypeKind::Long { endian: Endianness::Little, signed: true },
-///         adjustment: 0,
-///         endian: Endianness::Little,
-///     }))
-/// );
-///
-/// // Adjustment after closing paren
-/// assert_eq!(
-///     parse_offset("(0x3c.l)+4"),
-///     Ok(("", OffsetSpec::Indirect {
-///         base_offset: 0x3c,
-///         pointer_type: TypeKind::Long { endian: Endianness::Little, signed: true },
-///         adjustment: 4,
-///         endian: Endianness::Little,
-///     }))
-/// );
-/// ```
-///
-/// # Errors
-///
-/// Returns a nom parsing error if:
-/// - The input contains invalid number format (propagated from `parse_number`)
-/// - Input is empty or contains no parseable offset value
-/// - The offset value cannot be represented as a valid `i64`
-/// - Indirect offset has invalid pointer specifier or missing closing `)`
-pub fn parse_offset(input: &str) -> IResult<&str, OffsetSpec> {
-    let (input, _) = multispace0(input)?;
-
-    if input.starts_with('(') {
-        let (input, spec) = parse_indirect_offset(input)?;
-        let (input, _) = multispace0(input)?;
-        Ok((input, spec))
-    } else if let Some(rest) = input.strip_prefix('&')
-        && rest.starts_with('(')
-    {
-        // `&(...)`: relative wrapper around an indirect spec. Parse the
-        // inner indirect normally, then mark its result as relative so the
-        // evaluator adds it to the current anchor instead of treating it
-        // as an absolute file position. magic(5) uses this in rules like
-        // `&(&0.b+8)` to chain anchor-relative pointer reads.
-        let (rest, mut spec) = parse_indirect_offset(rest)?;
-        if let OffsetSpec::Indirect {
-            ref mut result_relative,
-            ..
-        } = spec
-        {
-            *result_relative = true;
-        }
-        let (rest, _) = multispace0(rest)?;
-        Ok((rest, spec))
-    } else if let Some(rest) = input.strip_prefix('&') {
-        // Relative offset: `&N`, `&+N`, or `&-N`. `parse_number` handles the
-        // bare and `-`-prefixed cases natively; `+` is consumed manually
-        // (see the indirect-offset adjustment parser for the same pattern).
-        let (rest, value) = if let Some(after_plus) = rest.strip_prefix('+') {
-            parse_number(after_plus)?
-        } else {
-            parse_number(rest)?
-        };
-        let (rest, _) = multispace0(rest)?;
-        Ok((rest, OffsetSpec::Relative(value)))
-    } else {
-        // Capture the leading `-` before `parse_number` consumes it: magic(5)
-        // `-0` means "0 bytes from the end of the file" -- the EOF position
-        // (`buffer.len()`), NOT absolute offset 0. Because `-0 == 0` in a
-        // signed integer the sign is otherwise lost, so detect it explicitly
-        // and encode `FromEnd(0)`. Used by e.g. gzip's `>>-0 offset >48` to
-        // gate the trailing-size trailer on the file being long enough.
-        // Other negative offsets (`-4`) keep their `Absolute(-4)` encoding,
-        // which the evaluator already resolves from the buffer end.
-        let starts_with_minus = input.starts_with('-');
-        let (input, offset_value) = parse_number(input)?;
-        let (input, _) = multispace0(input)?;
-        if starts_with_minus && offset_value == 0 {
-            return Ok((input, OffsetSpec::FromEnd(0)));
-        }
-        Ok((input, OffsetSpec::Absolute(offset_value)))
-    }
-}
-
-/// Parse comparison operators for magic rules
-///
-/// Supports both symbolic and text representations of operators:
-/// - `=` or `==` for equality
-/// - `!=` or `<>` for inequality
-/// - `<` for less-than
-/// - `>` for greater-than
-/// - `<=` for less-than-or-equal
-/// - `>=` for greater-than-or-equal
-/// - `&` for bitwise AND
-/// - `^` for bitwise XOR
-/// - `~` for bitwise NOT
-/// - `x` for any value (always matches)
-///
-/// # Examples
-///
-/// ```ignore
-/// use libmagic_rs::parser::grammar::parse_operator;
-/// use libmagic_rs::parser::ast::Operator;
-///
-/// assert_eq!(parse_operator("="), Ok(("", Operator::Equal)));
-/// assert_eq!(parse_operator("=="), Ok(("", Operator::Equal)));
-/// assert_eq!(parse_operator("!="), Ok(("", Operator::NotEqual)));
-/// assert_eq!(parse_operator("<>"), Ok(("", Operator::NotEqual)));
-/// assert_eq!(parse_operator("<"), Ok(("", Operator::LessThan)));
-/// assert_eq!(parse_operator(">"), Ok(("", Operator::GreaterThan)));
-/// assert_eq!(parse_operator("<="), Ok(("", Operator::LessEqual)));
-/// assert_eq!(parse_operator(">="), Ok(("", Operator::GreaterEqual)));
-/// assert_eq!(parse_operator("&"), Ok(("", Operator::BitwiseAnd)));
-/// assert_eq!(parse_operator("^"), Ok(("", Operator::BitwiseXor)));
-/// assert_eq!(parse_operator("~"), Ok(("", Operator::BitwiseNot)));
-/// assert_eq!(parse_operator("x"), Ok(("", Operator::AnyValue)));
-/// ```
-///
-/// # Errors
-///
-/// Returns a nom parsing error if:
-/// - Input does not start with a recognized operator symbol
-/// - Input is empty or contains no valid operator
-/// - Operator syntax is incomplete (e.g., just `!` without `=`)
-pub fn parse_operator(input: &str) -> IResult<&str, Operator> {
-    let (input, _) = multispace0(input)?;
-
-    let bytes = input.as_bytes();
-    let err = || nom::Err::Error(nom::error::Error::new(input, nom::error::ErrorKind::Tag));
-
-    // Dispatch on the first byte and inspect the second byte to choose between
-    // long-form and short-form operators. Boundary checks reject invalid
-    // sequences like "===", "&&", "^^", "~~", and "x42".
-    let (op, consumed) = match bytes.first().copied() {
-        Some(b'=') => {
-            // "=" or "==" -- reject "===" (and longer runs of '=')
-            if bytes.get(1).copied() == Some(b'=') {
-                if bytes.get(2).copied() == Some(b'=') {
-                    return Err(err());
-                }
-                (Operator::Equal, 2)
-            } else {
-                (Operator::Equal, 1)
-            }
-        }
-        // "!=" or bare "!" -- both map to NotEqual. magic(5) uses the bare
-        // form (e.g., `!0xb8c0078e` means "not equal to 0xb8c0078e"); the
-        // `!=` form is accepted as a convenience and matches operators in
-        // other parts of this parser.
-        Some(b'!') => {
-            if bytes.get(1).copied() == Some(b'=') {
-                (Operator::NotEqual, 2)
-            } else {
-                (Operator::NotEqual, 1)
-            }
-        }
-        Some(b'<') => {
-            // "<=", "<>", or bare "<"
-            match bytes.get(1).copied() {
-                Some(b'=') => (Operator::LessEqual, 2),
-                Some(b'>') => (Operator::NotEqual, 2),
-                _ => (Operator::LessThan, 1),
-            }
-        }
-        Some(b'>') => {
-            // ">=" or bare ">"
-            if bytes.get(1).copied() == Some(b'=') {
-                (Operator::GreaterEqual, 2)
-            } else {
-                (Operator::GreaterThan, 1)
-            }
-        }
-        Some(b'&') => {
-            // Reject "&&"
-            if bytes.get(1).copied() == Some(b'&') {
-                return Err(err());
-            }
-            (Operator::BitwiseAnd, 1)
-        }
-        Some(b'^') => {
-            // Reject "^^"
-            if bytes.get(1).copied() == Some(b'^') {
-                return Err(err());
-            }
-            (Operator::BitwiseXor, 1)
-        }
-        Some(b'~') => {
-            // Reject "~~"
-            if bytes.get(1).copied() == Some(b'~') {
-                return Err(err());
-            }
-            (Operator::BitwiseNot, 1)
-        }
-        Some(b'x') => {
-            // Word boundary: 'x' must not be followed by an alphanumeric or '_'
-            // (e.g., "x42" or "xfoo" is not AnyValue).
-            if input
-                .get(1..)
-                .is_some_and(|s| s.starts_with(|c: char| c.is_alphanumeric() || c == '_'))
-            {
-                return Err(err());
-            }
-            (Operator::AnyValue, 1)
-        }
-        _ => return Err(err()),
-    };
-
-    let remaining = &input[consumed..];
-    let (remaining, _) = multispace0(remaining)?;
-    Ok((remaining, op))
-}
-
-/// Parse the identifier operand of a `name` / `use` meta-type directive.
-///
-/// Called from [`parse_type_and_operator`] when the leading keyword is
-/// `name` or `use`. Enforces that the keyword is followed by whitespace,
-/// an identifier matching `[A-Za-z0-9_-]+`, and no further non-whitespace
-/// content on the line. Malformed identifiers such as `part2=foo`
-/// (operator-adjacent continuation) or `part 2` (split identifier) are
-/// rejected as parse errors rather than silently consumed as a message.
-fn parse_name_or_use_meta<'a>(
-    type_name: &str,
-    input: &'a str,
-) -> IResult<
-    &'a str,
-    (
-        TypeKind,
-        Option<Operator>,
-        Option<crate::parser::ast::ValueTransform>,
-    ),
-> {
-    use nom::character::complete::space1;
-
-    // Require at least one whitespace character between the keyword and
-    // the identifier. `space1` rejects an empty gap, which enforces
-    // "bare `name` / `use` with no identifier" as a parse error.
-    let (input, _) = space1(input)?;
-
-    // magic(5) allows a `\^` prefix on a `use` identifier to mean "invoke
-    // the named subroutine but flip the endianness of every read inside
-    // it" (libmagic `softmagic.c` `cvt_flip`). Consume the prefix and
-    // record it on `MetaType::Use::flip_endian` so the evaluator can apply
-    // the flip (issue #236). The `\^` prefix is meaningless on `name`
-    // declarations, so only `use` is inspected.
-    let (input, use_flip_endian) = if type_name == "use" {
-        input
-            .strip_prefix("\\^")
-            .map_or((input, false), |rest| (rest, true))
-    } else {
-        (input, false)
-    };
-
-    let (after_id, id) =
-        take_while(|c: char| c.is_alphanumeric() || c == '_' || c == '-').parse(input)?;
-    if id.is_empty() {
-        return Err(nom::Err::Error(NomError::new(
-            after_id,
-            nom::error::ErrorKind::AlphaNumeric,
-        )));
-    }
-
-    // The character immediately following the identifier must be
-    // whitespace or end-of-input. Anything else (e.g. `=`, `!`, `<`,
-    // `>`, `&`, `^`, `~`, `|`, punctuation) means `take_while` truncated
-    // a malformed identifier such as `part2=foo`: reject instead of
-    // silently treating the leftover text as a message.
-    if let Some(next_char) = after_id.chars().next()
-        && !matches!(next_char, ' ' | '\t' | '\n' | '\r')
-    {
-        return Err(nom::Err::Error(NomError::new(
-            after_id,
-            nom::error::ErrorKind::Alpha,
-        )));
-    }
-
-    // Handle the trailing text after the identifier. The two directives
-    // diverge here, matching GNU `file`:
-    //
-    // - `use`: magic(5) has no message slot for a `use` site, and GNU
-    //   `file` never emits a use-site's own description (verified: a
-    //   `use foo BAR` renders no `BAR`). Drop the trailing text up to
-    //   end-of-line so the caller parses an empty message.
-    // - `name`: the `name` line's OWN description IS emitted when the
-    //   subroutine is invoked via `use` -- Mach-O universal `0 name
-    //   mach-o \b [`, `0 name matlab4 Matlab v4 mat-file`, `0 name
-    //   algol_68 Algol 68 source text`, etc. PRESERVE the trailing text
-    //   so the caller's `parse_message` captures it as the rule message
-    //   (later stored in the name table by `extract_name_table` and
-    //   emitted at the `use` site). Dropping it here is what previously
-    //   made rmagic omit those fragments (e.g. the leading `[` of the
-    //   Mach-O universal bracket detail).
-    //
-    // We deliberately do NOT reject embedded whitespace inside the
-    // identifier itself (which would be a real malformed rule like
-    // `part 2`); that's enforced earlier when `take_while` truncates the
-    // identifier on the first non-id character.
-    let mut tail = after_id;
-    if type_name == "use" {
-        while let Some(rest) = tail.strip_prefix(' ').or_else(|| tail.strip_prefix('\t')) {
-            tail = rest;
-        }
-        if let Some(next_char) = tail.chars().next()
-            && !matches!(next_char, '\n' | '\r')
-        {
-            let line_end = tail.find(['\n', '\r']).unwrap_or(tail.len());
-            tail = &tail[line_end..];
-        }
-    }
-
-    let meta = if type_name == "name" {
-        MetaType::Name(id.to_string())
-    } else {
-        MetaType::Use {
-            name: id.to_string(),
-            flip_endian: use_flip_endian,
-        }
-    };
-    let (rest, _) = multispace0(tail)?;
-    Ok((rest, (TypeKind::Meta(meta), None, None)))
-}
-
-/// Parse a type specification with an optional attached bitwise-AND mask operator
-/// (e.g., `lelong&0xf0000000`).
-///
-/// Returns the `TypeKind`, an optional attached `Operator` (`&MASK`), and an
-/// optional pre-comparison `ValueTransform` (`+N`, `-N`, `*N`, `/N`, `%N`,
-/// `|N`, `^N`).
-///
-/// # Examples
-///
-/// ```ignore
-/// use libmagic_rs::parser::grammar::parse_type_and_operator;
-/// use libmagic_rs::parser::ast::{TypeKind, Operator, Endianness};
-///
-/// // Type without operator or transform
-/// let (_, (kind, op, transform)) = parse_type_and_operator("lelong").unwrap();
-/// assert_eq!(kind, TypeKind::Long { endian: Endianness::Little, signed: true });
-/// assert_eq!(op, None);
-/// assert_eq!(transform, None);
-///
-/// // Type with mask operator
-/// let (_, (kind, op, _)) = parse_type_and_operator("lelong&0xf0000000").unwrap();
-/// assert!(matches!(op, Some(Operator::BitwiseAndMask(_))));
-///
-/// // Type with arithmetic transform
-/// let (_, (kind, op, transform)) = parse_type_and_operator("lelong+1").unwrap();
-/// assert_eq!(op, None);
-/// assert!(transform.is_some());
-/// ```
-///
-/// # Errors
-/// Returns a nom parsing error if the input doesn't match the expected format
-pub fn parse_type_and_operator(
-    input: &str,
-) -> IResult<
-    &str,
-    (
-        TypeKind,
-        Option<Operator>,
-        Option<crate::parser::ast::ValueTransform>,
-    ),
-> {
-    use crate::parser::ast::{PStringLengthWidth, RegexCount, RegexFlags};
-
-    let (input, _) = multispace0(input)?;
-
-    let (mut input, type_name) = crate::parser::types::parse_type_keyword(input)?;
-
-    // `name` and `use` are meta-type directives with a mandatory
-    // identifier suffix. They short-circuit the operator/value parse
-    // path via `parse_name_or_use_meta`, which also rejects malformed
-    // identifiers (operator-adjacent continuations like `part2=foo` or
-    // split identifiers like `part 2`).
-    if type_name == "name" || type_name == "use" {
-        return parse_name_or_use_meta(type_name, input);
-    }
-
-    // Handle pstring suffixes: /B, /H, /h, /L, /l, and optional /J modifier
-    let mut pstring_length_width = PStringLengthWidth::OneByte;
-    let mut pstring_length_includes_itself = false;
-    if type_name == "pstring"
-        && let Some(suffix_rest) = input.strip_prefix('/')
-    {
-        let (rest, width, includes_j) = parse_pstring_suffix(suffix_rest)?;
-        input = rest;
-        pstring_length_width = width;
-        pstring_length_includes_itself = includes_j;
-    }
-
-    // Handle regex suffixes via the extracted helper. See
-    // `grammar/type_suffix.rs::parse_regex_suffix` for the full
-    // "any-order flag/count interleaving, duplicate counts rejected"
-    // semantics and the `RegexCount` collapse logic.
-    let mut regex_flags = RegexFlags::default();
-    let mut regex_count = RegexCount::Default;
-    if type_name == "regex"
-        && let Some(suffix_rest) = input.strip_prefix('/')
-    {
-        let (rest, (flags, count)) = parse_regex_suffix(input, suffix_rest)?;
-        regex_flags = flags;
-        regex_count = count;
-        input = rest;
-    }
-
-    // Handle search suffix. `search/N[/flags]` supplies an explicit scan
-    // window (e.g., `search/256`, `search/256/s`, `search/256/cs`); the
-    // count `N` is a `NonZeroUsize`, so `search/0` is rejected. A bare
-    // `search` (no `/` suffix, e.g. `>8 search /Count`) is ALSO accepted:
-    // its `range` stays `None`, meaning scan-to-end-of-buffer. magic(5)
-    // documents the count as required, but the reference `file` binary
-    // accepts the bare form (`str_range == 0`), so we follow the
-    // implementation rather than the spec. Disambiguation is unambiguous:
-    // a ranged suffix attaches `/` directly to the keyword, while a bare
-    // search is followed by whitespace then the value.
-    let mut search_range: Option<::std::num::NonZeroUsize> = None;
-    let mut search_flags = crate::parser::ast::SearchFlags::default();
-    if type_name == "search"
-        && let Some(suffix_rest) = input.strip_prefix('/')
-    {
-        let (rest, (range, flags)) = parse_search_suffix(input, suffix_rest)?;
-        search_range = Some(range);
-        search_flags = flags;
-        input = rest;
-    }
-
-    // Handle string flag suffixes (e.g., `string/w`, `string/cW`).
-    // magic(5) flag letters per libmagic `src/file.h`:
-    //   `/W` STRING_COMPACT_WHITESPACE
-    //   `/w` STRING_COMPACT_OPTIONAL_WHITESPACE
-    //   `/c` STRING_IGNORE_LOWERCASE  (pattern lowercase => file folded)
-    //   `/C` STRING_IGNORE_UPPERCASE  (pattern uppercase => file folded)
-    //   `/t` STRING_TEXTTEST          (text-mode hint)
-    //   `/T` STRING_TRIM              (trim pattern leading/trailing ws)
-    //   `/b` STRING_BINTEST           (binary-mode hint)
-    //   `/f` STRING_FULL_WORD         (post-match word-boundary check)
-    //
-    // `/B` is deliberately NOT accepted here -- it is the pstring
-    // 1-byte length-width letter (`CHAR_PSTRING_1_BE`) and is not a
-    // string flag in libmagic. An earlier draft of this loop
-    // accepted `'B'`; that was wrong and is now rejected. See
-    // GOTCHAS S6.6.
-    let mut string_flags = crate::parser::ast::StringFlags::default();
-    if type_name == "string"
-        && let Some(suffix_rest) = input.strip_prefix('/')
-    {
-        let mut consumed = 0usize;
-        for ch in suffix_rest.chars() {
-            match ch {
-                'W' => string_flags.compact_whitespace = true,
-                'w' => string_flags.compact_optional_whitespace = true,
-                'c' => string_flags.ignore_lowercase = true,
-                'C' => string_flags.ignore_uppercase = true,
-                't' => string_flags.text_test = true,
-                'T' => string_flags.trim = true,
-                'b' => string_flags.bin_test = true,
-                'f' => string_flags.full_word = true,
-                _ => break,
-            }
-            consumed += ch.len_utf8();
-        }
-        if consumed > 0 {
-            input = &suffix_rest[consumed..];
-        }
-        // If `consumed == 0`, the `/` is not followed by a known flag
-        // letter. Leave `input` pointing at the `/` so the value
-        // parser (or the trailing-junk check) fails meaningfully. This
-        // is the path that rejects `string/B` and `string/x`.
-    }
-
-    // Check for a pre-comparison value transform (e.g., `lelong+1` or
-    // `ulequad/1073741824`). magic(5) supports `+`, `-`, `*`, `/`, `%`,
-    // `|`, and `^` between the type keyword and the comparison value;
-    // the transform runs on the read value before the comparison
-    // operator and before printf-style format substitution.
-    let (input, value_transform) = parse_value_transform(input)?;
-
-    // Check for an attached bitwise operator with optional mask (e.g.,
-    // `&0xf0000000` or bare `&`). See `type_suffix::parse_attached_operator`
-    // for the recognized forms and their error behavior. magic(5) does
-    // not allow combining `&MASK` with another value transform on the
-    // same rule, so the parsers are sequential and either-or in
-    // practice.
-    let (input, attached_op) = parse_attached_operator(input)?;
-
-    let (input, _) = multispace0(input)?;
-
-    // Build Regex/Search directly from the parsed suffixes; fall back to
-    // `type_keyword_to_kind` for every other type. PString still uses the
-    // patch-after-construct pattern because `type_keyword_to_kind` supplies
-    // its `max_length` default and the suffix parser only produces the
-    // length-width and `/J` flag.
-    let type_kind = match type_name {
-        "regex" => TypeKind::Regex {
-            flags: regex_flags,
-            count: regex_count,
-        },
-        "search" => TypeKind::Search {
-            // `None` range = bare `search` = scan-to-EOF (see the suffix
-            // handling above); `Some(n)` = `search/N`.
-            range: search_range,
-            flags: search_flags,
-        },
-        _ => {
-            // `type_keyword_to_kind` returns:
-            //  * `Ok(Some(kind))` for every fully-specified keyword
-            //    (byte, short, long, quad, float/double, dates,
-            //    string, pstring and variants).
-            //  * `Ok(None)` for suffix-required keywords (`regex`,
-            //    `search`), which are handled by the match arms above
-            //    and should never reach this branch.
-            //  * `Err(UnknownTypeKeyword)` for a keyword that was never
-            //    produced by `parse_type_keyword`. Under the grammar's
-            //    normal flow this is unreachable because `type_name`
-            //    was just returned by `parse_type_keyword`, but the
-            //    function is `pub` and we do not rely on panics to
-            //    enforce the invariant -- we convert both "shouldn't
-            //    happen" cases into a nom parse error anchored at the
-            //    current input position so the parser can backtrack or
-            //    report a clean failure without aborting the process.
-            let Ok(Some(mut kind)) = crate::parser::types::type_keyword_to_kind(type_name) else {
-                return Err(nom::Err::Error(nom::error::Error::new(
-                    input,
-                    nom::error::ErrorKind::Tag,
-                )));
-            };
-            if let TypeKind::PString { max_length, .. } = kind {
-                kind = TypeKind::PString {
-                    max_length,
-                    length_width: pstring_length_width,
-                    length_includes_itself: pstring_length_includes_itself,
-                };
-            }
-            // Stamp the parsed string flags onto the `string` variant.
-            // `type_keyword_to_kind` returns `flags: StringFlags::default()`
-            // because the flag-bearing suffix is grammar-layer only; the
-            // type-keyword layer never sees it.
-            if let TypeKind::String { max_length, .. } = kind {
-                kind = TypeKind::String {
-                    max_length,
-                    flags: string_flags,
-                };
-            }
-            kind
-        }
-    };
-
-    Ok((input, (type_kind, attached_op, value_transform)))
-}
-
-/// Parse a type specification (byte, short, long, quad, string, etc.)
-///
-/// Supports various type formats found in magic files:
-/// - `byte` / `ubyte` - single byte (signed / unsigned)
-/// - `short` / `ushort` - 16-bit integer (native endian, signed / unsigned)
-/// - `leshort` / `uleshort` - 16-bit little-endian integer
-/// - `beshort` / `ubeshort` - 16-bit big-endian integer
-/// - `long` / `ulong` - 32-bit integer (native endian, signed / unsigned)
-/// - `lelong` / `ulelong` - 32-bit little-endian integer
-/// - `belong` / `ubelong` - 32-bit big-endian integer
-/// - `quad` / `uquad` - 64-bit integer (native endian, signed / unsigned)
-/// - `lequad` / `ulequad` - 64-bit little-endian integer
-/// - `bequad` / `ubequad` - 64-bit big-endian integer
-/// - `string` - null-terminated string
-/// - `pstring` - Pascal string (length-prefixed, supports `/B` (1-byte, default), `/H` or `/h` (2-byte), `/L` or `/l` (4-byte) suffixes)
-///
-/// # Examples
-///
-/// ```ignore
-/// use libmagic_rs::parser::grammar::parse_type;
-/// use libmagic_rs::parser::ast::{TypeKind, Endianness};
-///
-/// assert_eq!(parse_type("byte"), Ok(("", TypeKind::Byte { signed: true })));
-/// assert_eq!(parse_type("leshort"), Ok(("", TypeKind::Short { endian: Endianness::Little, signed: true })));
-/// assert_eq!(parse_type("bequad"), Ok(("", TypeKind::Quad { endian: Endianness::Big, signed: true })));
-/// assert_eq!(parse_type("string"), Ok(("", TypeKind::String { max_length: None, flags: StringFlags::default() })));
-/// ```
-///
-/// # Errors
-/// Returns a nom parsing error if the input doesn't match any known type
-#[allow(dead_code)] // Standalone helper exercised by grammar unit tests.
-pub fn parse_type(input: &str) -> IResult<&str, TypeKind> {
-    let (input, (type_kind, _, _)) = parse_type_and_operator(input)?;
-    Ok((input, type_kind))
-}
-
-/// Parse the indentation level and offset for magic rules
-///
-/// Handles both absolute offsets and hierarchical child rules with `>` prefix.
-/// Child rules can be nested multiple levels deep with multiple `>` characters.
-///
-/// # Examples
-///
-/// ```ignore
-/// use libmagic_rs::parser::grammar::parse_rule_offset;
-/// use libmagic_rs::parser::ast::OffsetSpec;
-///
-/// // Absolute offset
-/// assert_eq!(parse_rule_offset("0"), Ok(("", (0, OffsetSpec::Absolute(0)))));
-/// assert_eq!(parse_rule_offset("16"), Ok(("", (0, OffsetSpec::Absolute(16)))));
-///
-/// // Child rule (level 1)
-/// assert_eq!(parse_rule_offset(">4"), Ok(("", (1, OffsetSpec::Absolute(4)))));
-///
-/// // Nested child rule (level 2)
-/// assert_eq!(parse_rule_offset(">>8"), Ok(("", (2, OffsetSpec::Absolute(8)))));
-/// ```
-/// Parse rule offset with hierarchy level (> prefixes) and offset specification
-///
-/// # Errors
-/// Returns a nom parsing error if the input doesn't match the expected offset format
-pub fn parse_rule_offset(input: &str) -> IResult<&str, (u32, OffsetSpec)> {
-    let (input, _) = multispace0(input)?;
-
-    // Count the number of '>' characters for nesting level
-    let (input, level_chars) = many0(char('>')).parse(input)?;
-    let level = u32::try_from(level_chars.len()).unwrap_or(0);
-
-    // Parse the offset after the '>' characters
-    let (input, offset_spec) = parse_offset(input)?;
-
-    Ok((input, (level, offset_spec)))
-}
-
-/// Parse the message part of a magic rule
-///
-/// The message is everything after the value until the end of the line.
-/// It may contain format specifiers and can be empty.
-///
-/// # Examples
-///
-/// ```ignore
-/// use libmagic_rs::parser::grammar::parse_message;
-///
-/// assert_eq!(parse_message("ELF executable"), Ok(("", "ELF executable".to_string())));
-/// assert_eq!(parse_message(""), Ok(("", "".to_string())));
-/// assert_eq!(parse_message("  \tPDF document  "), Ok(("", "PDF document".to_string())));
-/// ```
-/// Parse the message/description part of a magic rule
-///
-/// # Errors
-/// Returns a nom parsing error if the input cannot be parsed as a message
-pub fn parse_message(input: &str) -> IResult<&str, String> {
-    let (input, _) = multispace0(input)?;
-
-    // Take everything until end of line, trimming whitespace
-    // Use take_while instead of take_while1 to handle empty messages
-    let (input, message_text) = take_while(|c: char| c != '\n' && c != '\r').parse(input)?;
-    let message = message_text.trim().to_string();
-
-    Ok((input, message))
-}
-
-/// Parse a strength directive (`!:strength` line)
-///
-/// Parses the `!:strength` directive that modifies rule strength.
-/// Format: `!:strength [+|-|*|/|=]N` or `!:strength N`
-///
-/// # Examples
-///
-/// ```ignore
-/// use libmagic_rs::parser::grammar::parse_strength_directive;
-/// use libmagic_rs::parser::ast::StrengthModifier;
-///
-/// assert_eq!(parse_strength_directive("!:strength +10"), Ok(("", StrengthModifier::Add(10))));
-/// assert_eq!(parse_strength_directive("!:strength -5"), Ok(("", StrengthModifier::Subtract(5))));
-/// assert_eq!(parse_strength_directive("!:strength *2"), Ok(("", StrengthModifier::Multiply(2))));
-/// assert_eq!(parse_strength_directive("!:strength /2"), Ok(("", StrengthModifier::Divide(2))));
-/// assert_eq!(parse_strength_directive("!:strength =50"), Ok(("", StrengthModifier::Set(50))));
-/// assert_eq!(parse_strength_directive("!:strength 50"), Ok(("", StrengthModifier::Set(50))));
-/// ```
-///
-/// # Errors
-///
-/// Returns a nom parsing error if:
-/// - Input doesn't start with `!:strength`
-/// - The modifier value cannot be parsed as a valid integer
-/// - The operator is invalid
-pub fn parse_strength_directive(input: &str) -> IResult<&str, StrengthModifier> {
-    // Helper to safely convert i64 to i32 with clamping to valid strength range.
-    // This prevents silent truncation to 0 on overflow while keeping values in bounds.
-    // Clamping to `[i32::MIN, i32::MAX]` is lossless via `as i32`, so no
-    // `unwrap()`/`expect()` is needed (AGENTS.md bans panic markers in
-    // library code regardless of whether the unwrap is provably safe).
-    #[allow(clippy::cast_possible_truncation)]
-    fn clamp_to_i32(n: i64) -> i32 {
-        n.clamp(i64::from(i32::MIN), i64::from(i32::MAX)) as i32
-    }
-
-    let (input, _) = multispace0(input)?;
-    let (input, _) = tag("!:strength")(input)?;
-    let (input, _) = multispace0(input)?;
-
-    // Parse the operator: +, -, *, /, = or bare number (implies =).
-    // Optional whitespace is permitted between the operator character and
-    // the operand to match GNU `file` magic(5) parsers, which accept forms
-    // like `!:strength / 2` in real-world magic files (e.g., the Minix
-    // entries in /usr/share/file/magic/filesystems).
-    //
-    // Use `preceded` + `Parser::map` (nom 8 idiom) rather than
-    // `map(pair(..), |(_, n)| ..)` so the throwaway `_` goes away and
-    // the composition matches the rest of this file's `.parse(input)?`
-    // style. Tuples implement `Parser` in nom 8 and are used as the first
-    // element of `preceded` to consume the operator char plus any trailing
-    // whitespace.
-    let (input, modifier) = alt((
-        // +N -> Add
-        preceded((char('+'), multispace0), parse_number)
-            .map(|n| StrengthModifier::Add(clamp_to_i32(n))),
-        // -N -> Subtract (parse_number handles negative directly; we
-        // need parse_decimal_number after the explicit `-` consumer
-        // so the sign is applied exactly once).
-        preceded((char('-'), multispace0), parse_decimal_number)
-            .map(|n| StrengthModifier::Subtract(clamp_to_i32(n))),
-        // *N -> Multiply
-        preceded((char('*'), multispace0), parse_number)
-            .map(|n| StrengthModifier::Multiply(clamp_to_i32(n))),
-        // /N -> Divide
-        preceded((char('/'), multispace0), parse_number)
-            .map(|n| StrengthModifier::Divide(clamp_to_i32(n))),
-        // =N -> Set
-        preceded((char('='), multispace0), parse_number)
-            .map(|n| StrengthModifier::Set(clamp_to_i32(n))),
-        // Bare number -> Set
-        parse_number.map(|n| StrengthModifier::Set(clamp_to_i32(n))),
-    ))
-    .parse(input)?;
-
-    Ok((input, modifier))
-}
-
-/// Check if a line is a strength directive (starts with !:strength)
-///
-/// # Examples
-///
-/// ```ignore
-/// use libmagic_rs::parser::grammar::is_strength_directive;
-///
-/// assert!(is_strength_directive("!:strength +10"));
-/// assert!(is_strength_directive("  !:strength -5"));
-/// assert!(!is_strength_directive("0 byte 1"));
-/// ```
-#[must_use]
-pub fn is_strength_directive(input: &str) -> bool {
-    input.trim().starts_with("!:strength")
-}
 
 /// Parse a complete magic rule line from text format
 ///
@@ -1255,252 +280,5 @@ pub fn parse_magic_rule(input: &str) -> IResult<&str, MagicRule> {
     Ok((input, rule))
 }
 
-/// Parse the comparison value for a string-family type.
-///
-/// libmagic never interprets a `string`/`pstring`/`string16`/`search`
-/// comparison value as a number: `0 string >0` compares against the
-/// literal ASCII byte `'0'` (0x30), and `>0.6.1` against the literal
-/// characters `0.6.1` -- not the integer 0 or the float 0.6. The generic
-/// [`parse_value`] tries its float and integer branches (see
-/// `value::parse_value`) before falling through, so a bareword like `0` or
-/// `0.6.1` was captured as `Value::Uint`/`Value::Float`. A subsequent
-/// comparison against the string field read from the file then yields no
-/// ordering (`String` vs `Uint`/`Float` is incomparable), so the rule
-/// silently never matched -- breaking real `>0` idioms such as
-/// `\b, name %s` / `face %s` / `palette %s` and version compares like
-/// `>0.6.1 ... version %s`.
-///
-/// This parser mirrors [`parse_value`]'s ordering for the two branches
-/// that are correct for string-family values -- a leading whitespace trim,
-/// then a quoted string (-> `Value::String`), then a hex/escape byte
-/// sequence (-> `Value::Bytes`, e.g. gzip's `\037\213` or `\177ELF`) -- but
-/// replaces the numeric (float/integer) branches with
-/// [`parse_bare_string_value`], so every remaining bareword resolves to a
-/// `Value::String`. The leading `multispace0` is load-bearing: it ensures
-/// the hex branch sees byte-identical input to what `parse_value` fed it,
-/// so an escape-heavy value cannot fall through to the lossy-UTF-8
-/// `parse_bare_string_value` path and corrupt a high byte (see the
-/// `high-byte-utf8-corruption-class` note).
-///
-/// Hex-*letter* barewords (`>AB`, `cafebabe`) still resolve to
-/// `Value::Bytes` via the unchanged hex branch, matching GOTCHAS S3.12 --
-/// only the numeric subset changes here.
-///
-/// # Errors
-/// Returns a nom parsing error only when the value is empty/whitespace-only
-/// (via [`parse_bare_string_value`]); quoted and hex forms are attempted
-/// first and never error out of this function on a non-empty token.
-fn parse_string_family_value(input: &str) -> IResult<&str, Value> {
-    // Trim leading whitespace up front so the hex branch below receives the
-    // same (trimmed) input `parse_value` would have handed it.
-    let (input, _) = multispace0(input)?;
-    if let Ok((rest, s)) = value::parse_quoted_string(input) {
-        return Ok((rest, Value::String(s)));
-    }
-    if let Ok((rest, bytes)) = value::parse_hex_bytes(input) {
-        return Ok((rest, Value::Bytes(bytes)));
-    }
-    parse_bare_string_value(input)
-}
-
-/// Parse a bare (unquoted) single-token string literal as a `Value::String`.
-///
-/// Used only as a fallback for string-family types (`string`, `pstring`,
-/// `regex`, `search`) when the strict [`parse_value`] alternatives all
-/// fail. Consumes leading whitespace, then reads a run of non-whitespace
-/// characters as the literal value, **interpreting magic(5) escape
-/// sequences** along the way: `\0`, `\n`, `\r`, `\t`, `\\`, `\"`, `\'`,
-/// `\NNN` (3-digit octal), and `\xNN` (hex). This supports magic-file
-/// rules like `0 string PNCIHISK\0 ...` where the trailing `\0` denotes
-/// a literal NUL byte that must be present in the file.
-///
-/// Without escape interpretation, the comparison value stored in the
-/// AST is the literal six-byte string `\` + `0` instead of `\x00`, and
-/// the rule never matches against a real on-disk byte sequence ending
-/// in NUL. This was a regression that prevented even simple top-level
-/// rules from matching when loaded from a magic file.
-///
-/// # Errors
-/// Returns a nom parsing error if the input contains no non-whitespace
-/// token (e.g. it is empty or consists entirely of whitespace).
-fn parse_bare_string_value(input: &str) -> IResult<&str, Value> {
-    let (input, _) = multispace0(input)?;
-    if input.is_empty() || input.starts_with(|c: char| c.is_whitespace() || c == '\n' || c == '\r')
-    {
-        return Err(nom::Err::Error(NomError::new(
-            input,
-            nom::error::ErrorKind::TakeWhile1,
-        )));
-    }
-
-    let mut bytes: Vec<u8> = Vec::new();
-    let mut remaining = input;
-    while let Some(ch) = remaining.chars().next() {
-        if ch.is_whitespace() || ch == '\n' || ch == '\r' {
-            break;
-        }
-        if ch == '\\' {
-            // Try a hex byte (`\xNN`) first since `parse_escape_sequence`
-            // doesn't recognise it.
-            if let Ok((rest, b)) = value::parse_hex_byte_with_prefix(remaining) {
-                bytes.push(b);
-                remaining = rest;
-                continue;
-            }
-            if let Ok((rest, esc)) = value::parse_escape_sequence(remaining) {
-                // `parse_escape_sequence` returns a `char`, but the
-                // escape table covers single-byte values (NUL, control
-                // chars, octal `\NNN` clamped to a `u8`). Cast back to
-                // `u8` so the buffer stays byte-accurate when the
-                // value is later compared against file bytes.
-                let code = esc as u32;
-                if let Ok(byte) = u8::try_from(code) {
-                    bytes.push(byte);
-                } else {
-                    // Escape produced a non-byte char (shouldn't happen
-                    // with the current escape grammar, but guard
-                    // anyway). Encode as UTF-8 so we never lose data
-                    // silently.
-                    let mut buf = [0u8; 4];
-                    bytes.extend_from_slice(esc.encode_utf8(&mut buf).as_bytes());
-                }
-                remaining = rest;
-                continue;
-            }
-            // Lone `\` followed by an unrecognised escape char: magic(5)
-            // getstr DROPS the backslash and keeps the character literally
-            // (`\<` -> `<`, `\^` -> `^`, `\ ` -> a literal space that
-            // continues the token because it lands in `bytes` rather than
-            // being re-examined by the whitespace-terminator check). This
-            // matches GNU `file`; the earlier "keep the backslash" behavior
-            // broke real rules like sgml's `0 string \<?xml\ version=`, which
-            // then never matched an actual `<?xml ...` document (XML files
-            // fell through to "ASCII text"). A genuine literal backslash is
-            // written `\\` and is already resolved by `parse_escape_sequence`
-            // above, so this branch only ever drops a backslash that was
-            // escaping a non-special character. A trailing lone `\` at
-            // end-of-input has no following char, so it stays literal.
-            if let Some(next) = remaining[1..].chars().next() {
-                let mut buf = [0u8; 4];
-                bytes.extend_from_slice(next.encode_utf8(&mut buf).as_bytes());
-                remaining = &remaining[1 + next.len_utf8()..];
-            } else {
-                bytes.push(b'\\');
-                remaining = &remaining[1..];
-            }
-            continue;
-        }
-        // Plain character: encode as UTF-8 (ASCII is one byte; non-ASCII
-        // is 2-4 bytes which matches how the file would store the same
-        // characters in a UTF-8 magic file).
-        let mut buf = [0u8; 4];
-        let utf8 = ch.encode_utf8(&mut buf).as_bytes();
-        bytes.extend_from_slice(utf8);
-        remaining = &remaining[ch.len_utf8()..];
-    }
-
-    if bytes.is_empty() {
-        return Err(nom::Err::Error(NomError::new(
-            input,
-            nom::error::ErrorKind::TakeWhile1,
-        )));
-    }
-
-    // Mirror `read_string_exact` (evaluator/types/string.rs): when the
-    // resolved bytes are valid UTF-8, return `Value::String` so `%s`
-    // output renders normally; when they are NOT -- e.g. a bareword like
-    // OS/2 INF's `HSP\x01\x9b\x00` (0x9b is invalid UTF-8), or an octal
-    // form like `AB\376` -- return the RAW bytes as `Value::Bytes`. A
-    // lossy `String` decode would turn 0x9b into U+FFFD (3 bytes 0xEF BF
-    // BD), which BOTH inflates the pattern's byte length (6 -> 8, so
-    // `read_string_exact` reads the wrong number of bytes) AND changes the
-    // byte value at that position, so the rule would silently never match.
-    // Cross-type `String`/`Bytes` equality and ordering (GOTCHAS S2.3)
-    // compare by byte sequence, so either variant compares correctly
-    // against the read value. This completes the read/parse symmetry:
-    // `read_string_exact` was fixed to return `Value::Bytes` on non-UTF-8
-    // slices, but the parse side kept lossy-decoding until this change.
-    match String::from_utf8(bytes) {
-        Ok(s) => Ok((remaining, Value::String(s))),
-        Err(e) => Ok((remaining, Value::Bytes(e.into_bytes()))),
-    }
-}
-
-/// Parse a comment line (starts with #)
-///
-/// Comments in magic files start with '#' and continue to the end of the line.
-/// This function consumes the entire comment line.
-///
-/// # Examples
-///
-/// ```ignore
-/// use libmagic_rs::parser::grammar::parse_comment;
-///
-/// assert_eq!(parse_comment("# This is a comment"), Ok(("", "This is a comment".to_string())));
-/// assert_eq!(parse_comment("#"), Ok(("", "".to_string())));
-/// ```
-/// Parse a comment line (starting with #)
-///
-/// # Errors
-/// Returns a nom parsing error if the input is not a valid comment
-pub fn parse_comment(input: &str) -> IResult<&str, String> {
-    let (input, _) = multispace0(input)?;
-    let (input, _) = char('#').parse(input)?;
-    let (input, comment_text) = take_while(|c: char| c != '\n' && c != '\r').parse(input)?;
-    let comment = comment_text.trim().to_string();
-    Ok((input, comment))
-}
-
-/// Check if a line is empty or contains only whitespace
-///
-/// # Examples
-///
-/// ```ignore
-/// use libmagic_rs::parser::grammar::is_empty_line;
-///
-/// assert!(is_empty_line(""));
-/// assert!(is_empty_line("   "));
-/// assert!(is_empty_line("\t\t"));
-/// assert!(!is_empty_line("0 byte 1"));
-/// ```
-#[must_use]
-pub fn is_empty_line(input: &str) -> bool {
-    input.trim().is_empty()
-}
-
-/// Check if a line is a comment (starts with #)
-///
-/// # Examples
-///
-/// ```ignore
-/// use libmagic_rs::parser::grammar::is_comment_line;
-///
-/// assert!(is_comment_line("# This is a comment"));
-/// assert!(is_comment_line("#"));
-/// assert!(is_comment_line("  # Indented comment"));
-/// assert!(!is_comment_line("0 byte 1"));
-/// ```
-#[must_use]
-pub fn is_comment_line(input: &str) -> bool {
-    input.trim().starts_with('#')
-}
-
-/// Check if a line ends with a continuation character (\)
-///
-/// Magic files support line continuation with backslash at the end of lines.
-///
-/// # Examples
-///
-/// ```ignore
-/// use libmagic_rs::parser::grammar::has_continuation;
-///
-/// assert!(has_continuation("0 string test \\"));
-/// assert!(has_continuation("message continues \\"));
-/// assert!(!has_continuation("0 string test"));
-/// ```
-#[must_use]
-pub fn has_continuation(input: &str) -> bool {
-    input.trim_end().ends_with('\\')
-}
 #[cfg(test)]
 mod tests;
